api/requirements.txt
fastapi==0.115.2
psycopg2-binary==2.9.10
uvicorn==0.32.0
